Move `NameFormatter`to production
